+++ sessionpool_test.go
